+++ internal/xds/utils/fractionalpercent/fractional_percent.go
